* commit 'b1fd56aba3a6eb97ea4723cbdde6cdb6565e4b02':
  更新进度，version

+++ fbi_base/src/com/fr/bi/stable/engine/index/TableIndexAdapter.java
@@ -50,7 +50,7 @@ public class TableIndexAdapter implements ICubeTableService {
     }
 
     @Override
-    public int getTableVersion(BIKey key) {
+    public long getTableVersion(BIKey key) {
         throw NULL_EXCEPTION;
     }
